# Introduce `r_wrap_or_clone()` and `r_wrap_or_copy()`