--- a/src/internal/encoding.c
+++ b/src/internal/encoding.c
@@ -131,7 +131,7 @@ r_obj* obj_attrib_encode_utf8(r_obj* x, r_obj* attrib) {
   }
   KEEP(attrib_new);
 
-  x = KEEP(r_clone_shared(x));
+  x = KEEP(r_wrap_or_clone_shared(x));
   r_poke_attrib(x, attrib_new);
 
   FREE(2);